SolesOfTheDamned
SolesOfTheDanmed is the passive ability script that goes onto gatherer. It has values you can change in the inspector. FlamesPerSecond which changes the amount of flames spawned per second, Duration which is how long each individual block of fire lasts from its creation so it disappears in trail form, damage which is how much damage each tick of fire does. You need to attach a fire prefab to the script. The fire prefab needs to have the FireForSolesOfTheDamned script on it.