Zone Momentum
The Zone Momentum when active and you are moving the character increases the acceleration and max speed of gatherer by the multiplier float every second in the zone momentum script. Once passive isn't active it resets the acceleration and maxspeed to the original. I capped the max speed and acceleration to 50 or gatherer literally becomes the flash. You can change the if statements in Zone Momentum to change the upper limits.